+++ src/FsCodec.NewtonsoftJson/Options.fs
@@ -5,13 +5,13 @@ open Newtonsoft.Json.Serialization
 open System
 open System.Runtime.InteropServices
 
+[<AbstractClass; Sealed>]
 type Options private () =
 
     static let defaultConverters : JsonConverter[] = [| OptionConverter() |]
 
     static let def = lazy Options.Create()
-
-    /// <summary>Analogous to <c>JsonSerializerOptions.Default</c> - allows for sharing/caching of the default profile as defined by <c>Options.Create()</c></summary>
+    /// <summary>Analogous to <c>System.Text.Json</c>'s <c>JsonSerializerOptions.Default</c> - allows for sharing/caching of the default profile as defined by <c>Options.Create()</c></summary>
     static member Default : JsonSerializerSettings = def.Value
 
     /// Creates a default set of serializer settings used by Json serialization. When used with no args, same as JsonSerializerSettings.CreateDefault()