+++ src/FsCodec.SystemTextJson/Options.fs
@@ -7,10 +7,10 @@ open System.Text.Json.Serialization
 
 #nowarn "44" // see IgnoreNullValues below
 
+[<AbstractClass; Sealed>]
 type Options private () =
 
     static let def = lazy Options.Create()
-
     /// <summary>Analogous to <c>JsonSerializerOptions.Default</c> - allows for sharing/caching of the default profile as defined by <c>Options.Create()</c></summary>
     static member Default : JsonSerializerOptions = def.Value